impl tratamento de erros da entidade veiculo

+++ src/main/java/com/api/parkingcontrol/services/VehicleServiceImpl.java
@@ -3,6 +3,8 @@ package com.api.parkingcontrol.services;
 import com.api.parkingcontrol.dtos.VehicleDTO;
 import com.api.parkingcontrol.models.Vehicle;
 import com.api.parkingcontrol.repositories.VehicleRepository;
+import com.api.parkingcontrol.services.exceptions.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
 import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
@@ -34,21 +36,29 @@ public class VehicleServiceImpl implements VehicleService{
 
     @Override
     public VehicleDTO findById(Long id) {
-        Vehicle entity = repository.findById(id).orElseThrow();
+        Vehicle entity = repository.findById(id).orElseThrow(()-> new ResourceNotFoundException(" Id not found !"));
         return new VehicleDTO(entity);
     }
 
     @Override
     public VehicleDTO update(Long id, VehicleDTO dto) {
-        Vehicle entity = repository.getReferenceById(id);
-        copyDtoToEntity(dto,entity);
-        entity.setUpdateDate(LocalDateTime.now());
-        entity = repository.save(entity);
-        return new VehicleDTO(entity);
+        try {
+            Vehicle entity = repository.getReferenceById(id);
+            copyDtoToEntity(dto,entity);
+            entity.setUpdateDate(LocalDateTime.now());
+            entity = repository.save(entity);
+            return new VehicleDTO(entity);
+        }
+        catch (EntityNotFoundException e){
+            throw new ResourceNotFoundException("Resource not found !");
+        }
     }
 
     @Override
     public void delete(Long id) {
+        if (!repository.existsById(id)){
+            throw new ResourceNotFoundException("Resource not found !");
+        }
         repository.deleteById(id);
     }
